+++ protocol/thrift/apache/adaptor/struct_codec.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package apache_adaptor
+package adaptor
 
 import (
 	"fmt"
